requirements.txt
python-docx>=1.2.0
mcp>=1.0.0
fastmcp>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
